Add `make_share_link` endpoint
Which makes a POST request to the branch api to get back a url shortlink for sharing a bagel. That said, this link doesn't pass validation because I don't have a correct validation_code. I then decompiled the android app to try to figure out how the validation code gets generated but was unsuccesful. So far now, I won't be able to automate the creation of the shared bagel link from Shehzam, but I can still redeem/dislike bagels on Shehzam's behalf and then open the app and go into history to share those bagels with Shehzad.